% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sm.R
\name{run_sm}
\alias{run_sm}
\title{Run SM test}
\usage{
run_sm(tree, data, trait, rep = 999)
}
\arguments{
\item{tree}{phylogeny to test}

\item{data}{dataframe of trait data, first column must contain tip labels of phylogeny}

\item{trait}{column name of trait of interest}

\item{rep}{number of randomisations to perform
run_sm()}
}
\description{
Run SM test
}

